A simple command-line application showcase.
This code creates a simple command-line application that allows you to add and list projects. You can customize this code as per your requirements and use it as a starting point for your portfolio.

The following tools are used in this Java command-line application:

Java Standard Library - The code uses various classes and methods from the Java Standard Library such as ArrayList, Scanner, System, and others.

Command-line Interface (CLI) - The application is a CLI application that interacts with the user through the console using text-based inputs and outputs.

Java Classes and Objects - The application uses classes and objects to create a portfolio of projects, where each project is represented as an object of the Project class. The Portfolio class manages a list of Project objects using an ArrayList.

Control Flow Structures - The application uses control flow structures such as if statements and switch statements to handle user inputs and navigate through different options.

These tools are commonly used in developing Java applications and provide a solid foundation for building a more complex portfolio in the future.

#Summary
The Java command-line application is a simple program that allows you to manage a portfolio of projects. It is designed to run in the console and interact with the user through a text-based interface.

The application consists of two main classes: Portfolio and Project. The Portfolio class is responsible for managing a list of Project objects, which represent individual projects in the portfolio. It provides the main functionality of the application through a series of methods, such as addProject and listProjects.

When the application is run, it displays a menu to the user with three options:

Add a project: Allows the user to enter the name and description of a new project and add it to the portfolio.

List all projects: Lists the name of all projects in the portfolio.

Quit: Exits the application.

The user interacts with the application by entering an option number from the menu, and the application responds by executing the corresponding functionality.

The Project class represents a single project in the portfolio. It contains two instance variables, name and description, which store the name and description of the project. The class also provides getter methods to retrieve the name and description of a project.

This application provides a basic template for a Java portfolio and can be customized or expanded as needed. For example, you could add additional functionality such as editing or deleting projects, or adding more detailed information about each project.